+++ apps/modernjs-ssr/remote/modern.config.ts
@@ -1,0 +1,19 @@
+import { appTools, defineConfig } from '@modern-js/app-tools';
+import { moduleFederationPlugin } from '@module-federation/modern-js';
+
+// https://modernjs.dev/en/configure/app/usage
+export default defineConfig({
+  dev: {
+    port: 3051,
+  },
+  runtime: {
+    router: true,
+  },
+
+  plugins: [appTools(), moduleFederationPlugin()],
+  server: {
+    ssr: {
+      mode: 'stream',
+    },
+  },
+});