+++ packages/dts-plugin/src/plugins/DevPlugin.ts
@@ -142,12 +142,18 @@ export class DevPlugin implements WebpackPluginInstance {
       const TEMP_DIR = path.join(`${process.cwd()}/node_modules`, BasicTempDir);
       const filepath = path.join(TEMP_DIR, `live-reload.js`);
 
-      DevPlugin.ensureLiveReloadEntry({ name }, filepath);
-      compiler.hooks.afterPlugins.tap('MFDevPlugin', () => {
-        new compiler.webpack.EntryPlugin(compiler.context, filepath, {
-          name,
-        }).apply(compiler);
-      });
+      if (typeof compiler.options.entry === 'object') {
+        DevPlugin.ensureLiveReloadEntry({ name }, filepath);
+        Object.keys(compiler.options.entry).forEach((entry) => {
+          const normalizedEntry = compiler.options.entry[entry];
+          if (
+            typeof normalizedEntry === 'object' &&
+            Array.isArray(normalizedEntry.import)
+          ) {
+            normalizedEntry.import.unshift(filepath);
+          }
+        });
+      }
     }
 
     const defaultGenerateTypes = { compileInChildProcess: true };