+++ packages/dts-plugin/vite.config.ts
@@ -1,15 +1,15 @@
 /// <reference types="vitest" />
 import { defineConfig } from 'vite';
+import path from 'path';
 
 import { nxViteTsPaths } from '@nx/vite/plugins/nx-tsconfig-paths.plugin';
 
 export default defineConfig({
-  cacheDir: '../../node_modules/.vite/dts-plugin',
+  cacheDir: '../../node_modules/.vitest/dts-plugin',
 
   plugins: [nxViteTsPaths()],
 
   test: {
-    // cache: false,
     cache: {
       dir: '../../node_modules/.vitest',
     },
@@ -20,5 +20,6 @@ export default defineConfig({
     ],
     reporters: ['default'],
     testTimeout: 60000,
+    setupFiles: [path.resolve(__dirname, './tests/setup.ts')],
   },
 });