+++ packages/bridge/bridge-shared/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @module-federation/bridge-shared
 
+## 0.3.1
+
+## 0.3.0
+
 ## 0.2.8
 
 ## 0.2.7