+++ packages/third-party-dts-extractor/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @module-federation/third-party-dts-extractor
 
+## 0.3.1
+
+## 0.3.0
+
 ## 0.2.8
 
 ## 0.2.7